app/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "telefix"
version = "0.0.1"
description = "Yaserviceru is a bot application"
readme = "README.md"
authors = [{name = "Oleg Souzdalev", email = "odsuzdalev@gmail.com"}]
# homepage = "https://github.com/osuzdalev/yaserviceru"
requires-python = "~=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Programming Language :: Python :: 3.9"
]
dependencies = [
    "aiohttp",
    "aiosignal",
    "anyio",
    "APScheduler",
    "argcomplete",
    "astroid",
    "async-timeout",
    "attrs",
    "Authlib",
    "black",
    "build",
    "cachetools",
    "certifi",
    "cffi",
    "charset-normalizer",
    "click",
    "cryptography",
    "decorator",
    "dill",
    "docutils",
    "exceptiongroup",
    "filelock",
    "frozenlist",
    "fsspec",
    "h11",
    "httpcore",
    "httpx",
    "huggingface-hub",
    "idna",
    "importlib-metadata",
    "isort",
    "jaraco.classes",
    "Jinja2",
    "joblib",
    "keyring",
    "loguru",
    "markdown-it-py",
    "MarkupSafe",
    "mccabe",
    "mdurl",
    "more-itertools",
    "mpmath",
    "multidict",
    "mypy-extensions",
    "networkx",
    "nh3",
    "nltk",
    "numpy",
    "openai",
    "package-name",
    "packaging",
    "pandas",
    "pathspec",
    "Pillow",
    "pkginfo",
    "platformdirs",
    "psycopg",
    "pyaes",
    "pyasn1",
    "pycparser",
    "pydantic",
    "Pygments",
    "pylint",
    "pyproject_hooks",
    "python-dateutil",
    "python-dotenv",
    "python-telegram-bot",
    "pytz",
    "PyYAML",
    "readme-renderer",
    "regex",
    "requests",
    "requests-toolbelt",
    "rfc3986",
    "rich",
    "rsa",
    "ruff",
    "safetensors",
    "scikit-learn",
    "scipy",
    "sentence-transformers",
    "sentencepiece",
    "six",
    "sniffio",
    "sympy",
    "Telethon",
    "threadpoolctl",
    "tiktoken",
    "tokenizers",
    "tomli",
    "tomlkit",
    "torch",
    "torchvision",
    "tqdm",
    "transformers",
    "twine",
    "typing_extensions",
    "tzlocal",
    "urllib3",
    "validators",
    "weaviate-client",
    "xmltodict",
    "yarl",
    "yq",
    "zipp"
]

[project.scripts]
telefix = "telefix.telefix:main"

[tool.setuptools]
packages = ["telefix"]